Rephrased wording around unique choice.
I reordered the paragraphs so that type theory is not cut in half by set theory.
I tried to simplify the explanation of what happens in set theory.
It is irrelevant that in *formal* set theory we never construct anything, because in informal set theory we do, and mathematicians know this. We should compare informal to informal.