Revert "polish the cover and title page"
This reverts commit e69e2f4721d5bfa8b928923e630b005997d36001.

Could you please not mess with the cover page, because it takes me 30 minutes to
get the fonts and the size and the spacing done every time. There is a miniscule
advantage in using the \outtitle macros. I am MUCH happier to have the text hard-coded
because every space counts, and with the macros I lose control over what the cover
looks like. Thanks.

Conflicts:
	front.tex
	main.tex