Removed the \ourtitle macros, again.
These macros have a miniscule advantage, if any. It is not clear from
their names where they appear, so we have to search anyway. They are not
useful for the title page, and \title and \author do not not appear visible
anyhow. So please stop putting them in because I will not use them on the
title page. And it's not like we are going to change the title a dozen
times.