Moved macros to `macros.tex`, closing #155.
The rationale for this is that someone had to look at what all
the macros were doing anyway, so I did. The reason all macros
are in one place is that this way people will discover pre-existing
macros. I am not sure anymore that placing them all in `macros.tex`
was such a brilliant idea, but it probably does no great harm.