Cleaning up after changes by @awodey.
We are in a phase where every change in the use of macros must
be considered carefully because it easily creates inconsistencies.
I am guilty of doing this with the `\P` macro, for example.
Therefore, if you are going to change the use of any macro (in
this case it was `\indidone` and `\indidtwo`) then you *must*:

* check all uses of it (there were some in `formal.tex`),
* fix the symbols index

Also, do *not* leave old text in as LaTeX comment. This:

* screws with the git diffs and makes it harder to see what changes
* increases the probability that an unwanted new paragraph will apear