Conforming to US Trade format.
@awodey please check that you're ok with this. I had to make
the title of the first part shorter in order to fit it on
a single line in the US Trade format. Also, in that format
we need to use two columns instead of three. There is a bit
of an unfortunate page break towards the end.

Generate the US Trade format with `make hott-ustrade.pdf`