Rename inequality to disequality, closing issue #226
Note that we use "inequality" to refer to the order of
cardinal numbers, reals, and surreals, in the later chapters.
Therefore we should not use "inequality" here to mean "not equal".
We opt for "disequality" because some communities already use
that word, and nobody seems to use "unequality" systematically.
In any case, the use of "disequality" is limited to just a
tiny section in the first chapter, so this does not really matter.