Makefile: stop always checking for warnings from "makeindex"
The warning messages seem mostly spurious.  For example, if we
have \indexdef{x} and \index{x} end up on the same page, then in
the index we'll have the page number mentioned twice, once in bold
and once not in bold.  Since repagination might change things,
it's probably best not to "fix" it.