Add .travis.yml for a Travis-CI instance
Someone with admin bits for this repo will need to go to
https://travis-ci.org/profile and sign in with github and enable Travis
for this repo.

If Travis is enabled, new commits and pull requests get auto-built by
Travis.  If someone is feeling ambitious, it's possible to make Travis
do other tasks, like auto-uploading a new version of the book somewhere,
or parsing `git blame` to update the eratta numbers and pushing the
appropriate eratta labels.