Add exercise about finite choice
I'm not sure if I should include a note about the exercise requiring
functional extensionality, or about the exercise not requiring the
assumptions that A(x) are sets and P(x, a) are mere propositions.  I'm
also not sure if I should be marking an erratum for this.

This closes #622.