Add dvimode, remove pagecolor package
This closes #666.

Travis will now test dvi mode for us.  This is primarily useful for
htlatex.

Move the bmpsize hack for bounding boxes to a separate .tex file, to be
used in other files.

We also now silence the first two runs of (pdf)latex, because travis was
complaining about the log file being too big.