small grammar correction (between vs. among)
Although "between" can be used to compare more than 2 things in some specific cases, I don't think this sentence qualifies as one of them. Furthermore, when I first read the sentence, I kept wondering if it was meant a relationship between the $n$-types and some other thing (which should be inferred from context) or among the $n$-types themselves. I suppose that it was meant the later.