Close #1101 with "they"
Even though @andrejbauer stated that the vote would be open until Sunday, April 17th, we have decided unanimously that we need not wait until then, as there is a clear preference for the 'they' option, which we shall implement promptly.

Adoption of a suitable Code of Conduct and related topics may still be discussed in Issue #1111.