Unvendor packages to fix CI
Something related to "fixcleverref" is incompatible with
texlive-full=2024.20250309-1 on CI or Ubuntu 25.04. The CLI had
"Latex failed to resolve 216 citation(s)". The end of the introduction
had "??" instead of links, and the word "section" in unstyled text
before headings.

Unvendoring all the (now-) unnecessarily vendored local
packages fixes the problem. I enabled `-g` for `latexmk` to fix the .pdf
not being rebuilt when I delete it.

Fixes: #1172